services/forecasting/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.0
pandas==2.2.3
numpy==1.26.4
prophet==1.1.5
pymongo==4.10.1
python-dotenv==1.0.1

